+++ tests/closed.test.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2020 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import 'mocha';
+import 'babel-polyfill';
+
+import * as firebase from '@firebase/rules-unit-testing';
+
+import { readRulesFile, getAuthedDb } from './util';
+
+describe('[Closed Rules]', () => {
+  before(async () => {
+    await firebase.loadFirestoreRules({
+      projectId: 'closed-rules',
+      rules: readRulesFile('../rules/closed.rules')
+    });
+  });
+
+  it('should deny a read any any path to closed rules', async () => {
+    const db = getAuthedDb('closed-rules', undefined);
+
+    await firebase.assertFails(
+      db
+        .collection('any')
+        .doc('doc')
+        .get()
+    );
+  });
+});
