+++ tsconfig.json
@@ -23,6 +23,6 @@
         ]
     },
     "include": [
-        "src/**/*"
+        "tests/**/*"
     ]
 }
